[WIP] build: Set up Android build to use RN from Git.
This comes from just following steps 2, 3, 4 in upstream's instructions:
  https://facebook.github.io/react-native/docs/building-from-source.html

Note also the `local.properties` and other local setup in our guide:
  https://github.com/zulip/zulip-mobile/blob/master/docs/howto/rn-from-git.md

The major way this isn't yet fully baked is that before running
`react-native run-android` (and after running `yarn`), you have to
fix up the Gradle configurations of our dependencies, like so:

  sed -i 's/compile .com.facebook.react:react-native:[0-9.]*+./compile project('\'':ReactAndroid'\'')/' \
    node_modules/{@remobile/,}*/android/build.gradle

Before we can merge this, we need to fix that, of course.  The job
this is doing is a lot like what step 4 in upstream's guide is
described as being supposed to do... but apparently that step isn't
enough.

The right fix will be some other change to our Gradle config; perhaps
one that amounts to making the same substitution as my sed command
above.  Gradle configs are just code (and with a pretty decent API,
too), so in the end it's always possible to do what we need.  Some
debugging, documentation-reading, and experimenting will be required
to find a good way to do this.