ZulipStatusBar: Resolve a baffling couple of Flow fixmes.
The error we were getting here were like this:

  Cannot create ZulipStatusBar element because ST [1]
  is not a React component.

     src/main/MainScreenWithTabs.js
      18│
      19│     return (
      20│       <View style={[styles.flexed, styles.backgroundColor]}>
      21│         <ZulipStatusBar />
      22│         <OfflineNotice />
      23│         <MainTabs />
      24│       </View>

     flow-typed/npm/react-redux_v5.x.x.js
 [1] 119│     ST: {[_: $Keys<Com>]: any}

That `ST` is in the libdef for `connect`; it's supposed to represent
the static properties of the component class itself.  I think it's
beside the point of the error, a consequence of Flow finding a
contradiction and then deriving arbitrary results.

When I deleted that bit from the signature in the libdef (that
`flow-typed/...` file mentioned in the error), I got this rather more
meaningful error message:

  Cannot get props.narrow because property narrow
  is missing in props [1].

     src/common/ZulipStatusBar.js
     70│ export default connect((state: GlobalState, props) => ({
     71│   safeAreaInsets: getSession(state).safeAreaInsets,
     72│   theme: getSettings(state).theme,
     73│   backgroundColor: props.backgroundColor || getTitleBackgroundColor(props.narrow)(state),
     74│   orientation: getSession(state).orientation,
     75│ }))(ZulipStatusBar);
     76│

     src/main/MainScreenWithTabs.js
 [1] 21│         <ZulipStatusBar />

That still doesn't seem right to me!  If it's missing, OK, the result
is `undefined`... and what we do with that is precisely to pass it
someplace that's happy with `undefined`.

But at least it's comprehensible; also, easy to reassure with a bit of
explicitness, which never hurt anybody.  So fine, be explicit.