wip set vibration
This pattern roughly approximates the Zulip notification sound.

Also for >=26 we need to configure on the channel, not the notif.
TODO: This needs to go with rotating the channel ID for existing installs.

For API <26, the corresponding change will also depend on dropping
`builder.setDefaults`, which requests the default vibration as well as
sound and light behavior.  (For API >= 26, the `defaults` value on the
notification is ignored along with the sound, vibration, and lights
settings it would override.)