wip serverKnownToken... see TODO next steps in body
Really the data layout I think I want is:

 * serverKnownToken belongs in `accounts` store, not `realm`;
   it doesn't expire on switching the active account.
   (`ackedToken`?)

 * `token` i.e. the latest locally-learned token goes
   somewhere global and non-persistent, so probably `session`.

Then on each launch we get `token` from the system; and for every
logged-in account that doesn't have a matching acked-token, give it
the news.  And the listener for (rare) GCM token changes just pings
JS to do that same flow again.

This means `Account` will diverge from `Auth`, as we've long
anticipated in principle but haven't seen a concrete reason to do.
Did a big refactor F 12-28 of the `accounts` selectors, in order
to straighten those out some so as to be better prepared for such
a change.