wip messy failed attempt to type redux-persist-migrate
The tough obstacle here is to give an appropriate type for Action.
The type should be "either RehydrateAction, or else an object where
`type` is *not* `REHYDRATE`"... and I think there there may not
actually be a way to write that generically, i.e. short of listing
out every possible value for the `type` property.

Now that we have this library vendored, we can just do that, by
importing something like our `Action` type.  That'll probably be the
way we end up typing this file.

But it's not clear what an actual library like this one should do.
Most likely the answer involves a $FlowFixMe in the implementation...
or perhaps using TypeScript instead.  (I don't know whether TypeScript
would offer a better solution here.)

Another complication visible here is that although I can add things to
`redux-persist/constants` in this libdef file and import them just
fine, I can't do the same for `redux-persist`.  This one is more
baffling -- it's either an outright bug in Flow, or something silly
that I'm completely missing -- but also easy to work around by just
putting the `RehydrateAction` in our own JS code rather than a libdef.
The `REHYDRATEE` etc. is part of trying to debug that; the
`RehydrateAction` is what we'd actually want here from such a libdef.