wip compose box: Factor out closeComposeMenu.
This makes it more explicit what's going on and why in all these
places.

There's no effect on behavior, so long as these are all inside
React event handlers:
  https://stackoverflow.com/a/48610973/458193

... but hmm, they're not -- see calls to `.handle...` in
`set*InputValue`, which in turn are called in
componentWillReceiveProps as well as in some more event handlers.

... OK, but that's only in the `editMessage` case which is rare;
and the only possible bad effect is performance, so meh in that case.