wip react-redux: Find a way to make `connect` actually catch selector/props type mismatches.
Note that getOwnUser has the wrong type -- which should cause a type
error when used here on OwnAccountCard.

I was surprised to find today that it didn't!  Even in this very
straightforward scenario where the selector type is completely wrong,
the Props type is exact, there's no use of props in mapStateToProps,
and there's no local state.

On some investigation, here's a lead on a way to make it work.
Needs some work to get back to the generality of the existing libdef
(reflecting the wildly dynamic semantics of `connect`)... but it is
at least fully *generic* in that everything it knows about the specific
components, it learns from parameters.

I'm also pleased with the technique this demonstrates for developing
(esp. for iterating on) such a libdef type, which I hadn't hit on
before.