android,autocomplete:Fix text doesn't autocompletes on selecting item.
This got introduced in 041ce55e7, where background color is set to the
parent of `absolute` autocomplete view.

This seems to be a RN issue that, `absolute` positioned components
doesn't receive touchable events if inside component which has
background color set. (will try to verify it on new RN project)

So set background color to compose input view, sibling of autocomplete
view instead of parent of autocomplete view.

Fixes: #3378