wip narrow: Add streamNameFromNarrow, and demonstrate its use.
TODO: have first commit throw in an isStreamNarrow to make pure refactor;
  then ditto the getAllTopicsForStream commit;
  then remove both those isStreamNarrow calls, fixing this bug.

This can replace a lot of the places where we have fragile and opaque
expressions like `narrow[0].operand`, relying on guard conditions
elsewhere in the code like `isStreamOrTopicNarrow`.

Here, also fix an "ActiveFoo" name, which hasn't been accurate since
we stopped getting the narrow from nav state here a long time ago.
This function can just take a stream name directly; call it
`fetchTopicsByStreamName`.

This is nearly a pure refactor.  The attentive reader will notice one
wrinkle: `fetchTopicsByStreamName` will now get farther (going ahead
and fetching) if its caller `handleTopicFocus` is ever called with a
topic narrow, where previously it was a no-op.  Can that ever happen?
Well, it can get called if `getCanSelectTopic`... and that is true on
a topic narrow if `editMessage` is true.

Which means this actually fixes a bug!  Previously, on message edit we
would neglect to fetch the topics, and so topic autocomplete wouldn't
work.