stream list: Cut disused `selected` feature.
This hasn't been used since this lived in the nav drawer, long
ago.  As a result it's suffered some known bitrot -- note the
type `boolean | string`, which makes no sense -- and is likely
to have suffered some silent bitrot in the styles.

If we bring back something like this in some kind of nav drawer,
we can re-add something like this then.  We won't want to use
this verbatim anyway, because (in addition to the bogus type,
which would need fixing) it's unlikely the styles will still be
quite what we want.

(Of course this will always be here in the history to be
consulted, as much as we like.)