StreamList: Remove disused `selected` prop.
This bit of code was quite confused and confusing, setting the default
value for this prop to a boolean `false` but then testing it against
each stream's name, a string.  That left it hard to understand what
the prop might be intended to mean.

The recent commit aada8313c removed the one place where this prop was
used (some code which had itself been dead for well over a year), so
now just take it out.

It's possible in the future that we'll want functionality like this
again.  In that case we can easily add it back, and do so more cleanly.