StreamItem: Cut disused prop `isSelected`.
In the parent commit, we removed the one place this prop was referred
to; as discussed there, that codepath hadn't actually been used in
well over a year.

It's possible that at some point in the future we might want
functionality like this again.  But in that case it'll be easy to add
it back -- even to borrow the exact code from before this commit, if
we so desire, thanks to version control.

In the meantime, this case contributes a good chunk of the complexity
in the logic controlling colors in this component.  That logic is
rather tangled and badly in need of some straightening-out; and the
more cases it needs to handle, the more work it will be to do that.
So, we take out the complexity that isn't serving any purpose in order
to help with clarifying the remaining complexity that corresponds to
what the code is really accomplishing.