style: Cut a spurious StyleSheet.flatten, and associated goo.
When passed a style object, this is just the identity function.
Its only useful effect is to take an array of such objects,
and flatten them into one.

So, replace our one use of `StyleSheet.flatten` with just a
direct use of its argument, which is always an object.

That in turn allows us to cut some awkward goo that just hedged
against the impossible possibility that `flatten` would return
null or undefined.