redux: Add place for recent private conversations to Redux store.
This commit introduces the reducers required to introduce this data
into the Redux state, during `doInitialFetch`. It only catches the
`REALM_INIT` action at the moment, and the rest of the actions will be
added in a later commit with details.

It does not yet store the data received from the server in REALM_INIT.
Some additional work will need to be done to convert the data received
from older server versions.

[ray: Split original commit into two parts, this being the second.
  Temporarily disabled REALM_INIT reducer, just in case.]