wait for REALM_INIT to show UI after login... and try debugging
With this change, there's a wacky behavior where after logging in and
the loading screen, we get... the auth screen.

In the console with redux-logger, plus these ad-hoc log lines:

 * REALM_INIT, INITIAL_FETCH_COMPLETE, and COMPLETE_TRANSITION happen,
   the nav state is just [main]...

 * and then RealmScreen#componentDidMount gets called for some reason,
   with initialRealm of the realm we just logged into.

 * That executes cheerfully through to "did navigateToAuth", which
   comes after a NAVIGATE that gives us [main, auth].  Until then it
   was still [main].

 * Then "Warning: Can't perform a React state update on an unmounted
   component."  Must be the one in `finally`; but this also indicates
   that as of the top of `tryRealm` (where there's also a `setState`),
   this component was mounted.

It's quite mysterious to me why a `RealmScreen` seems to be mounted
when we haven't navigated to it.  Not sure how next to debug, either.