Add debugging info that pinpoints the cause of the issue.
This shows "null" on the new JS engine, and "undefined" when
remote-debugging JS.  (On iOS, where we have no such native module.)

And I haven't yet confirmed this, but it must be that it will
show "undefined" on the old JS engine, too -- because that is
what we look for in ZulipAsyncStorage.  Oof.