wip deps: Add @zulip/shared, for code shared from the webapp.
TODO: replace file: with normal version number (once uploaded)

The initial version of the shared package provides one module
`typing_status`, which implements the sender-side state machine for
typing notifications.

For this commit, we don't yet switch over to using that for real --
there's some nontrivial code required, so we leave that for a separate
upcoming commit.  Do include an import and a no-op call to the module,
the better to end-to-end confirm that the toolchain is all working.