wip ios: attempt getting push notifs in dev
Doesn't actually work, though. :-/  Not sure why.

The aim here was to get Cmd-r in Xcode to produce a build that could
receive notifications, by having it sign everything with production
keys.

For this to do anything useful, you'll want to be using the "scheme"
called "ZulipMobile release-mode", added in the parent commit.  That
way Cmd-r will build with the Release configuration.

Resources consulted include:

  General:
  https://developer.apple.com/documentation/usernotifications/registering_your_app_with_apns

  On the `aps-environment` "entitlement":
  https://developer.apple.com/documentation/bundleresources/entitlements/aps-environment

  General:
  https://developer.apple.com/documentation/usernotifications/setting_up_a_remote_notification_server/sending_notification_requests_to_apns
  That one says in part:

    Use the production server for your shipping apps and the
    development server for testing.

Also went and created a new "provisioning profile" at:
  https://developer.apple.com/account/resources/profiles
which is the one named here.  Previously we didn't have any that
hadn't expired -- thanks to the "Automatically manage signing"
feature, we hadn't needed them.  (We *did* and do have a perfectly
good *certificate*.  That profile pulls in the cert, and I guess
the "Automatically manage" feature takes the place of profiles in
pulling in an appropriate cert.)