wip ZulipStatusBar [nfc]: un-nest StatusBar
AFAICT this component doesn't actually care what its parent's size,
background, etc. etc. are -- because it doesn't render anything at its
actual spot in the tree.  Rather, it just merges some properties onto
the app's (or root view's or something's?) global idea of how the
status bar should be.

So, sticking it inside this View does nothing but give a misleading
idea of how this works.

See:
  https://facebook.github.io/react-native/docs/statusbar

For background on the iOS side (which I suspect is the primary
influence on the design of that RN API), see:
  https://developer.apple.com/design/human-interface-guidelines/ios/bars/status-bars/
  https://developer.apple.com/documentation/uikit/uiviewcontroller/1621416-preferredstatusbarstyle
and the various other UIViewController properties linked in "See Also"
of the latter.

For background on the Android side, see:
  https://developer.android.com/training/system-ui
  https://developer.android.com/reference/android/view/Window.html#setStatusBarColor(int)
  https://developer.android.com/reference/android/view/View.html#SYSTEM_UI_FLAG_FULLSCREEN