wip ZulipStatusBar: Set basic properties even in landscape.
This should fix an issue on Android: when in landscape, the actual
status bar still shows, but we neglect to set its background color.

None of these props we pull out of the PORTRAIT conditional should,
IIUC the docs, affect whether the status bar is actually shown in the
first place.