wip TODO notes on insets and status bar
I'm no longer confident that RN's SafeAreaView is going to be that
helpful for us.

From studying its implementation (which is quite small -- tiny, if you
tune out the bits that are for iOS <=10 only), what it does is
 * makes a boring view containing its children
 * and *sets the padding* (for RN's layout engine -- Yoga? -- to see,
   I believe) to whatever UIKit says its safeAreaInsets are.

That's cool but it leaves open the question of how to control what
bleeds to the visible-yet-obscured areas covered by those insets.
In particular,
 * when we have a brightly-colored nav bar, we really want to extend
   that color behind the status bar
 * also it would be nice to have e.g. the compose box's background
   extend behind the home indicator area at the bottom.

Because SafeAreaView just amounts to setting the padding, it
implicitly leaves whatever's in the component's background to show
through there.  We can certainly set the background... but that covers
*all* sides.  No good when what we want is the colorful nav-bar color
(from the stream color) extending only to the top.

None of this is hard, just so long as we wire up the information about
what the insets should be.

(One point of comparison that I found helpful in thinking about this
was Flutter:
  https://api.flutter.dev/flutter/widgets/SafeArea-class.html
  https://github.com/flutter/flutter/blob/1aedbb1835/packages/flutter/lib/src/widgets/safe_area.dart
)

Still, the architecture of it shouldn't be like the status quo --
random single-direction offsets thrown around in random places.
Instead we should have... a container component, something like
SafeAreaView, but with a few more knobs to say what background color
to extend in which direction.

Actually, better than that (and simpler!): just like SafeAreaView, but
*with some boolean flags* (all default-true) to control which sides to
add the padding on.  Then a caller could e.g., for ChatScreen,
 * have a safe-area for top only, with the nav bar color;
 * inside that, one for bottom only, with the compose box color;
 * inside that, one for left and right only, with back to the app's
   main (theme-dependent) background color.

Or perhaps in other words: the version of this container that has
knobs for what background color goes where is easy to implement on top
of the one with just boolean flags saying what directions to handle at
all.  One advantage of the simpler version is that the caller gets
quite explicit control of the question of what to do at corners:
e.g. does the top left get the "top" background or the "left"
background.

Hmm also handy: in landscape mode on an iPhone X relative, there are
insets at sides and bottom but not top.  We might want the nav bar
color to extend to left and right.  To do that, we'd want to have the
nav bar *outside* any safe-area container that covers the sides... and
the side-covering container would instead go inside the nav bar, and
also inside each of its vertically-laid-out siblings (including unread
notice etc.), so that their respective backgrounds would extend to the
sides.

How to implement that improved, more-controllable safe-area container?
Kind of seems like the *right* way is probably some Swift or Obj-C
code -- imitating RN's SafeAreaView, particularly for the plumbing
with the RN layout system (which I have only the vaguest inkling of),
but just accepting a few more boolean props.  That shouldn't be much
code, right?

A more proximately-easy alternative would be to go further down the
path of using data maintained in Redux and based on listening in JS
for updates to the safe-area insets.  That might even be the best for
at least initial experimentation, and fixing some of the most
conspicuous bugs.

---

BTW a tab dump of some pages I had open when studying this:
  https://facebook.github.io/react-native/docs/safeareaview
  https://developer.apple.com/documentation/uikit/uiview/positioning_content_relative_to_the_safe_area
  https://developer.apple.com/documentation/uikit/uiview/2891103-safeareainsets
  https://developer.apple.com/documentation/uikit/uiview/2891104-safeareainsetsdidchange?language=objc
  https://medium.com/rosberryapps/ios-safe-area-ca10e919526f

  https://developer.apple.com/documentation/uikit/uiview
  https://developer.apple.com/documentation/uikit/uiview/1622482-layoutsubviews
  https://developer.apple.com/documentation/uikit/uiview/1622507-layoutifneeded
  https://developer.apple.com/library/archive/documentation/WindowsViews/Conceptual/ViewPG_iPhoneOS/Introduction/Introduction.html#//apple_ref/doc/uid/TP40009503