wip msglist html [nfc]: encapsulate computing PM narrow string
TODO update comment -- much of this landed earlier

The "filter out self" logic doesn't actually go away here, because
it had been serving two different purposes!  We were using it both
for

 * computing the canonical string identifier to name the PM narrow
   internally, in order to navigate there if the user taps the
   header; and

 * computing the list of names to present in the header to identify
   the conversation to the user.

It happens that we use the same somewhat wrinkled logic in both
cases.  But that's kind of a coincidence -- in fact for the internal
canonical string identifier we might be better off dropping that
logic, and either always listing all the recipients (including self)
or always listing only the non-self recipients (so that the self-1:1
thread is represented with an empty list of (non-self) recipients.)

So, instead of using the same implementation here for both
computations, entrust the computation of the narrow's canonical
string identifier to the central narrows code.

Also add a comment explaining this somewhat wrinkled logic for its
remaining purpose.