tests [nfc]: Set up mocks for `logging` module.
We'll want to use this in the next commit, where we add some
`logging.warn` calls in code that has test coverage: it'll allow us
to (a) suppress those warnings from spewing when the tests are run,
plus (b) check in the test that we really do call `logging.warn`.

Meanwhile, it has no effect on tests that don't specifically do
something with these mocks -- which at this stage is all of our
tests.

By my reading of the Jest docs, we shouldn't need this here in our
global setup file -- we should be able to just call `jest.mock` in
the respective test files where we actually want the mocking.  But I
found that that didn't work for me; such calls just had no effect,
and both test and application code got the real, unmocked module.
That'd be good to figure out and fix in general, but fortunately
this particular module is both small and of broad interest.