and push sendTyping* all the way to CleanNarrow!
We will need to make this dirtier again in the case of old servers.

Hmm.  Not sure how I'd best like to handle that.

The most "convert at the edge" approach would be to indeed pass
CleanNarrow, and then right here consult the state if needed to
convert to emails.