outbox [nfc]: Send based on `display_recipient`, not `narrow`.
This was the last place we were using the `narrow` property on
outbox messages, so taking it out will mean we can remove that
property in the next commit.  Which is good because (a) the `Narrow`
type is a great big mess, and the fewer different parts of the code
are using it, the fewer need to be migrated to a cleaned-up version
of it; and (b) it was never an appropriate type to use to describe a
single message's destination.

I don't love the "display recipient" values either, but their
structure is at least less wild than Narrow, and their semantic
range is an appropriate match for "where this individual message
is sent".  And whenever we decide it's that type's turn to get
straightened out, this usage should fit in nicely to be swept up
with all the others, because it comes right after a check of the
display_recipient's friend the `type: 'stream' | 'private'` field.

This commit makes a small change in the data we send to the server
for sending a PM: compared to the old code's behavior, this code
will always differ by including one more email at the end of the
list, namely the self-user's email.  (To see that that's exactly
what changes, look at `mapEmailsToUsers`, which produces the
`display_recipient` value.)  The comment in the code explains why
that causes no functional change.

That comment, in turn, is based on reading both the code as of
zulip/zulip@97d7d31b6 (in 2013-03), the same commit referred to in
our e4a83be68, and also as of current master (zulip/zulip@fa1059aa2).
The relevant logic lives today in get_recipient_from_user_profiles,
and is essentially unchanged from the 2013 version.

(We could have arranged to send the server exactly the same data as
the old code, by slicing off the last element of `display_recipient`.
But that feels like it'd be more brittle, relying on what's really
an implementation detail of `mapEmailsToUsers`.)