search: Update `messages` before clearing `isFetching`.
This change has no behavioral effect if the fetch throws an
exception; if the fetch succeeds, its effect is to swap the order
in which we call `setState` to update `messages` and `isFetching`.

That swap, in turn, has no effect if React batches the two calls,
because the updates commute.  So in a future React version where
`setState` calls are batched by default (perhaps React 17):
  https://stackoverflow.com/a/48610973/378130
  https://reactjs.org/docs/react-component.html#setstate
this change will be NFC.  (That SO answer is by Dan Abramov and is
effectively part of the React docs.)

Seen as an NFC change, this is a small cleanup: it simplifies
tracing what happens to `messages` by making it `const`, and
consuming it promptly.

(The setting and clearing of `isFetching` become separated by a few
more lines of code.  But they have a nice, dual structure to them
which the two halves of the `messages` handling didn't, paired off
with the help of `try .. finally`.)

As long as the order can matter, though, the order that updates
`messages` first is the right one.  If `render` is called with
`isFetching` back to false but `messages` still at its old value,
then we'll present the old data as if it's the result of having
completed the search.  Much better, even briefly, to have a "still
loading" indicator linger but already show the new results.  (And
in the current UI implemented in SearchMessagesCard, even that
doesn't happen: after this change, the intermediate state renders
identically either to before both updates or to after them.)