wip upgrades: eslint
TODO:
 * Bunch of new lint errors, as usual; review, and fix or disable.
 * prettier-eslint seems to have totally broken.  Fix.

This is preparation to upgrade to babel-eslint v11-beta, which
requires eslint >=6.

Done with:

  $ yarn add -D eslint-plugin-react-hooks@^1.7.0
  $ yarn upgrade --latest $(
      tools/deps \
        | grep eslint \
        | grep -ve eslint-plugin-jest -e eslint-plugin-react-hooks
    )
  $ yarn yarn-deduplicate && yarn

The eslint-plugin-react-hooks bit is because the newer
eslint-config-airbnb -- the one compatible with ESLint 6 -- has a
peer-dependency on it.

The exclusion of eslint-plugin-jest is because of a different
peer-dependency that one of its dependencies brings in.