ios: Start using CocoaPods again.
There was an attempt to use CocoaPods in aded466, 2017-03-19.

That was removed in 4096e71, 2017-07-18.

Some cleanup was done in 861744b, 2020-02-05.

To paraphrase Greg on #3983:

CocoaPods is a tool for managing iOS libraries. Many people use it
in their React Native apps, but there some tricky bits that have
impeded universal adoption.

RN v0.60 signals a full embrace of the use of CocoaPods by making it
work more smoothly with RN. CocoaPods becomes required in v0.61.

So, use CocoaPods, writing a Podfile based on the example Podfile
given for react-native v0.59, at
https://github.com/facebook/react-native-website/blob/ded79d2cf4456d8b1a4f67c2cdc1391789e70617/docs/integration-with-existing-apps.md.
(That doc isn't live on the react-native docs website because of
https://github.com/facebook/react-native-website/issues/1603.)

Originally, we were going to start using CocoaPods as part of the RN
v0.60 upgrade, but we have a fresh reason to want to use CocoaPods,
with its own deadline: #3964, Apple auth. Someone has made a handy
Expo package to handle Apple auth, but we can't use Expo packages
without either fully committing to using the Expo SDK (a dramatic
step) or using a package called react-native-unimodules, which lets
you select individual Expo packages. react-native-unimodules
requires the use of CocoaPods.

Fixes: #3983