webview build: Build to an intermediate staging directory for Android.
... instead of stepping on the AGP's toes and hoping its dance card is
empty tonight.

The assets-merge task provided by the Android Gradle Plugin builds to
an undocumented output directory. Unfortunately, we'd been writing to
that same destination directory; this is not safe if done before the
assets-merge task runs.

Up until now, we hadn't noticed because the assets-merge had normally
been a no-op for us -- we have no asset directories declared in our
Android Studio project, so nothing for the merge task to merge. On the
first run, though, it's _not_ a no-op: it'll wipe the output directory
as it sets up the other directories in ${buildDir}/intermediates/,
even though it does nothing thereto thereafter.

We'd originally been writing directly to that directory in mimicry of
react-native-vector-icons, which does the same; but r-n-v-i does some
additional violence to the task-dependency graph in order to run
*after* the assets-merge task, rather than before it.

We could follow its lead there, but it's much simpler for us just to
work with the AGP's assets-merge system rather than work around it. We
instead build to a temporary asset-directory dynamically added in the
Gradle project.

Fixes #4018.