usersReducer: Handle `EVENT_USER_UPDATE` a little bit.
I'm wary of suddenly starting to clobber any arbitrary property in
users, so I'm starting small with just `avatar_url`. I chose this in
particular because we're about to use our own new data type here,
the `AvatarURL`.

Now, when someone updates their avatar, it will be live-updated in
`state.users`; see discussion, where Tim said to check that this is
being done [1].

Note that it's not yet updated in `state.messages`, so the message
list won't be live-updated when someone changes their avatar. See
discussion [2] on implementing this.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/user_avatar_url_field_optional/near/909056.
    In retrospect, I think he actually meant to make sure the server
    was behaving correctly -- but we might as well start doing the
    right thing here too.
[2] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/user_avatar_url_field_optional/near/992469