BackNavigationHandler: Delete, as handled naturally.
Some notable commits in the history of this logic:

0688b9b7f Handle Android hardware back button press. (#1125)
6785245e6 Improve and fix Back button on Android
ef99027f5 navigation: Split AppWithNavigation

Nothing suggests that we've been trying to do anything more
complicated than the following default behavior of React Navigation
(in the assumed absence of react-navigation-redux-helpers, which we
removed in a recent commit) [1]:

"""
By default, when user presses the Android hardware back button,
react-navigation will pop a screen or exit the app if there are no
screens to pop. This is a sensible default behavior, but there are
situations when you might want to implement custom handling.
"""

Logic like what we remove here was probably necessary when using
react-navigation-redux-helpers; that package gives an example [2]
that looks suspiciously similar to what we've been using. So, to be
safe, we don't remove until here, after we've definitely stopped
using r-n-r-h.

[1] https://reactnavigation.org/docs/4.x/custom-android-back-button-handling/
[2] https://github.com/react-navigation/redux-helpers#back-button