ZulipMobile: Set initial route declaratively; remove "initial navigation".
In a recent commit, we ensured that none of React Navigation's
components get mounted until the store has rehydrated. The data from
rehydration is just what we need to tell React Navigation what the
"initial route" is. With that data, we can remove the awkward
`InitialNavigationDispatcher` logic, which would be difficult to
coordinate with the `NavigationService`'s initialization schedule in
React Navigation 5 [1].

So, instead of doing an "initial navigation", tell React Navigation
from its first appearance what its initial route should be. In React
Navigation 4, this means not calling `createStackNavigator` until we
know what to pass for `initialRouteName` and `initialRouteParams`
there (with those being computed from data from the now-rehydrated
Redux store). Take care not to call `createStackNavigator` on every
`render`; instead, do it once, in `ZulipAppContainer`'s constructor,
and store it in an instance property.

The mechanics of initializing the root navigator will be simpler
with React Navigation v5's component-based API. It'll be normal to
just pass `initialRouteName` as a prop when we create a
`<Stack.Navigator />`, i.e., `initialRouteName` doesn't have to be
baked in to the navigator from its beginning (which, here, we do by
passing it when we call `createStackNavigator`).

We don't build `ZulipAppContainer` to be a drop-in replacement for
`AppContainer`, with which `NavigationService` can interface from
the outside. Rather, we encapsulate the connection with
`NavigationService` inside `ZulipAppContainer`. This is really an
accident of what's easiest to do; I didn't find `React.forwardRef`
flexible enough to expose `AppContainer`'s `ref` as
`ZulipAppContainer`'s `ref`, while still accomplishing the above
(calling `createStackNavigator` once, at the right time, and storing
the result in a sensible way).

[1] https://reactnavigation.org/docs/navigating-without-navigation-prop/#handling-initialization