i18n: Offer zh_TW instead of zh-Hant.
The existing zh-Hant translation appears to use Taiwanese vocabulary,
so it seems like it was effectively a zh_TW translation in the first
place under another name.  Between the two of them, the zh_TW
translation is the one that's now nearly complete, so prefer that one.
See chat discussion:
  https://chat.zulip.org/#narrow/stream/58-translation/topic/zh-Hant.20.2F.20zh_TW/near/1044288

There are several ways to name the different zh* translations.
We follow Wikipedia's choices.  Facebook does the same (except with
ASCII parens rather than full-width parens).  Apple and Google write
zh_TW as "繁體中文（台灣）", i.e. leading with "繁體" which identifies
the traditional script.