i18n: Sort languages in picker UI by endonym, not English name.
This makes a lot more sense when you're using any language
other than English, and it's what most software seems to do.

The other defensible alternative is to sort by the language code.
In order to make sense, that requires also displaying the code
in the UI.  This is what Wikipedia does in its preferences UI
on the web:
  https://en.wikipedia.org/wiki/Special:Preferences
but I'm not sure I've ever seen it anywhere else.