js [nfc]: Rewrite Array#forEach calls into for..of loops.
The most interesting part of this is probably the tools used to do it:
a codemod script added in this commit, which uses `recast` to do its
work.  Then we run it, using the handy `jscodeshift` CLI:

  $ npx jscodeshift -t tools/codemod/foreach-to-for-of.js src/ \
    && git restore src/third/ src/webview/js/ \
    && tools/test lint prettier --fix

We leave `src/third/` unmodified because that's by definition code
that we're generally leaving in the style of an upstream version.

We leave `src/webview/js/` unmodified because that's code we have
Babel compile for us in webviews, potentially running fairly old
browser versions -- and empirically, if we introduce `for..of`
there, the output code becomes more complex.

As noted in the script's jsdoc, the major way this could make an
incorrect transformation is if the value we're calling a `forEach`
method on isn't actually an Array, and the method's behavior isn't
actually equivalent to for..of.  I've read through the diff, and I
believe it is in fact an Array in every case.

As mentioned in a comment, this code draws heavily from the commit
message on Anders's zulip/zulip@02511bff1.  All the `recast` and
`ast-types` code, which is what actually inspects the AST and edits
it, is based on that example.  The handy CLI around the outside is
based on a previous experiment with jscodeshift, in this directory.