outbox [nfc]: Make explicit the odd "async" loop in trySendMessages.
This loop doesn't do what it looks like it does.  It looks like a
loop that tries to send one message, awaits that, then tries the
next, and so on.  In fact, because Array#forEach does nothing with
the return values of its callback -- in particular it doesn't notice
if they're promises or anything else, and doesn't await them -- the
actual behavior is to fire off a bunch of requests in parallel, wait
for nothing, and ignore errors.

That behavior isn't good, and we should fix it; that's #3881.

For a start, just make the code more explicit about what it actually
does: desugar the async/await keywords to the equivalent use of
Promise#then, and add a comment for good measure.

This also smooths the way to rewriting Array#forEach here as part of
an automated sweep across our code.