wip do webview js after all -- just use loose mode
  $ npx jscodeshift -t tools/codemod/foreach-to-for-of.js src/ \
    && git restore src/third/ \
    && tools/test lint prettier --fix
    && tools/generate-webview-js

The case where this can do something wrong is if (a) the iterable
isn't an Array, and (b) in particular it has some cleanup it wants
to do at the end, and (c) the loop body ends the loop early, with a
`break` or `return` or `throw`.  In that case, with loose mode, the
cleanup won't get run as it should.

In our code, I believe we never actually iterate through anything
other than an Array.  In particular we don't have any fancy
iterators with complex logic that'd require cleanup at the end of
the loop.  So go ahead and use loose mode for this transform, which
reportedly avoids a substantial slowdown.

This also matches what metro-react-native-babel-preset gives us for
our main codebase running in the RN environment.