wip NOMERGE enable React profiling, with a tiny ad-hoc Babel plugin
This makes React.Profiler work in a production build.  Upstream
doesn't describe a way to do this with RN, but they do for react-dom:

  https://fb.me/react-profiling
  aka https://gist.github.com/bvaughn/25e6233ae

and we follow the same approach here -- which is to rewrite some
imports within the React implementation so that they get a "profiling"
version of some key internals, instead of the "prod" version.

I initially tried using babel-plugin-module-resolver for this, because
someone reported that working for them.  It didn't; and after some
debugging, it turns out that approach stopped working in RN v0.61:
that plugin only rewrites absolute imports, never relative imports,
and it happens that some of the key imports we need to rewrite were
changed from absolute to relative in RN v0.61 (at commit
facebook/react-native@69d1ed731.)

The thing we actually want here is very specific, a little odd, and
doesn't match the intended use case of that plugin or necessarily any
other plugin off the shelf; so after debugging this enough to
understand why that wasn't working, we know enough that it's simplest
to just have our own tiny plugin specific to this purpose.

The one reason this is NOMERGE is that it's not conditional -- it
applies to all production builds.  Could be made mergeable with just a
bit more code in babel.config.js, to make it opt-in.