ChatScreen: Simplify `componentDidUpdate` and fix a bug.
Greg points out that this code has been unnecessarily tricky [1]:

"""
The old code at this spot is unnecessarily tricky, as I stare at it
again: instead of this conditional, it could just unconditionally
set `this.shouldFetchWhenNextFocused = true`.

Then the code just below here would see that, and if `isFocused`
were indeed true, it'd go and do a fetch, and set
`this.shouldFetchWhenNextFocused` back to false. That'd be simpler
-- and also closer to what the hooks version ends up looking like.
So if you do that as a prep commit, then it'll be a simplification
in itself and also make it easier to reason about the refactor to
hooks.
"""

He also notices a bug that this fixes:

"""
If should-fetch is already true, and then this gets called with
a(nother) change to `eventQueueId` at a time when we're focused...
then the existing code will go and fetch, and *leave* should-fetch
true. So then if we get unfocused and focused again, we'll fetch a
second time, even if the queue hasn't changed.
"""

[1] https://github.com/zulip/zulip-mobile/pull/4300#discussion_r523303810