wip get a name for a thunk action
Our Redux perf logging logs the names of thunk actions.  But typically
the thunk actions are anonymous functions -- they're just produced by
action creators that have names.  This makes it sometimes mysterious
where some time is going.

Experiment with giving one frequent thunk action a name.

This is a little ugly in the code, but it does work:

  Dispatching:          >fetchMessagesInNarrow
  Dispatching:            >_callee
  Dispatch sub-time:    1ms MESSAGE_FETCH_START reducer
  Dispatch time:    12ms     MESSAGE_FETCH_START
  Dispatch time:   20ms   _callee
  Dispatch time:   20ms fetchMessagesInNarrow

We might try to do it systematically despite the ugliness, or to do it
selectively and try to cover thunk actions that can take significant
time in the logs.